+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - More automated tests to improve robustness (https://github.com/riparias/gbif-alert/issues/93, https://github.com/riparias/gbif-alert/issues/131)
 - Remove deprecated settings (https://github.com/riparias/gbif-alert/issues/265)
 - The application is now available in Dutch
+- Fix an old and annoying bug for administrators: https://github.com/riparias/gbif-alert/issues/146
+- New data import script (specific to the LIFE RIPARIAS instance)
 
 # v1.5.0 (2023-09-26)